pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "otelmini"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.9"
license = "Apache-2.0"
keywords = []
authors = [
  { name = "Pablo Collins", email = "pabcolli@cisco.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "opentelemetry-api",
]

[project.urls]
Documentation = "https://github.com/pmcollins/otelmini#readme"
Issues = "https://github.com/pmcollins/otelmini/issues"
Source = "https://github.com/pmcollins/otelmini"

[project.scripts]
otel = "otelmini.distro:auto_instrument"

[tool.hatch.version]
path = "src/otelmini/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "oteltest",
  "pytest",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/otelmini tests}"

[tool.coverage.run]
source_pkgs = ["otelmini", "tests"]
branch = true
parallel = true
omit = [
  "src/otelmini/__about__.py",
]

[tool.coverage.paths]
otelmini = ["src/otelmini", "*/otelmini/src/otelmini"]
tests = ["tests", "*/otelmini/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.pytest.ini_options]
markers = ["slow: slow running tests (any test that doesn't feel instantaneous)"]
testpaths = ["tests"]
pythonpath = ["."]

[project.entry-points.opentelemetry_traces_exporter]
otlp = "otelmini.trace:HttpSpanExporter"
otlp_json_http = "otelmini.trace:HttpSpanExporter"
console = "otelmini.trace:ConsoleSpanExporter"

[project.entry-points.opentelemetry_metrics_exporter]
otlp = "otelmini.metric:HttpMetricExporter"
otlp_json_http = "otelmini.metric:HttpMetricExporter"
console = "otelmini.metric:ConsoleMetricExporter"

[project.entry-points.opentelemetry_logs_exporter]
otlp = "otelmini.log:HttpLogExporter"
otlp_json_http = "otelmini.log:HttpLogExporter"
console = "otelmini.log:ConsoleLogExporter"

[tool.hatch.envs.hatch-static-analysis]
config-path = "none"

[tool.ruff]
exclude = ["tests/"]
lint.ignore = [
  "UP007",    # Use X | Y for type union
  "FA100",    # Missing from __future__ import annotations
  "EM101",    # Exception string literal
  "EM102",    # Exception f-string literal
  "W293",     # Blank line contains whitespace
  "ARG002",   # Unused method argument
  "PLW0603",  # Global statement (used for singleton pattern)
  "PLW2901",  # Loop variable overwritten
  "TRY003",   # Long exception messages
  "BLE001",   # Catching blind Exception (intentional for instrumentation)
]
